+++ source/backend/plugin/DssiPlugin.cpp
@@ -1265,7 +1265,7 @@ public:
 
                     const EngineMidiEvent& engineEvent(event.midi);
 
-                    uint8_t status  = MIDI_GET_STATUS_FROM_DATA(engineEvent.data);
+                    uint8_t status  = uint8_t(MIDI_GET_STATUS_FROM_DATA(engineEvent.data));
                     uint8_t channel = event.channel;
 
                     // Fix bad note-off (per DSSI spec)
